Split up app.ts
