+++ ts/viewmodels/bag.ts
@@ -1,0 +1,15 @@
+import * as ko from "knockout";
+import * as _ from "underscore";
+import { Game } from "../scrabble/game";
+
+export class Bag {
+    public remaining: KnockoutComputed<Array<[string, number]>>;
+
+    public constructor(game: Game) {
+        this.remaining = ko.pureComputed(() => {
+            var bag = game.currentStatus().bag.sort();
+            // Returns key/value pairs as tuples [string, number]
+            return _.pairs(_.countBy(bag, _.identity));
+        });
+    }
+}
